+++ tiledb/sm/cpp_api/query.h
@@ -1077,6 +1077,32 @@ class Query {
   }
 
   /**
+   * Set the query config.
+   *
+   * Setting configuration with this function overrides the following
+   * Query-level parameters only:
+   *
+   * - `sm.memory_budget`
+   * - `sm.memory_budget_var`
+   * - `sm.sub_partitioner_memory_budget`
+   * - `sm.var_offsets.mode`
+   * - `sm.var_offsets.extra_element`
+   * - `sm.var_offsets.bitsize`
+   * - `sm.check_coord_dups`
+   * - `sm.check_coord_oob`
+   * - `sm.check_global_order`
+   * - `sm.dedup_coords`
+   */
+  Query& set_config(const Config& config) {
+    auto ctx = ctx_.get();
+
+    ctx.handle_error(tiledb_query_set_config(
+        ctx.ptr().get(), query_.get(), config.ptr().get()));
+
+    return *this;
+  }
+
+  /**
    * Set the coordinate buffer.
    *
    * The coordinate buffer has been deprecated. Set the coordinates for